requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-ftp==3.0.0
apache-airflow-providers-google==8.1.0
apache-airflow-providers-http==3.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-postgres==5.0.0
apache-airflow-providers-sqlite==3.0.0
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Babel==2.10.3
blinker==1.4
cachelib==0.9.0
cachetools==4.2.2
Cartopy @ git+https://github.com/SciTools/cartopy.git@357bd6b4c775e80b8721d065c0990fe97ee5f138
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.1
cftime==1.6.1
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
cron-descriptor==1.2.30
croniter==1.3.5
cryptography==36.0.2
cycler==0.11.0
db-dtypes==1.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
eventlet==0.33.1
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.37.1
future==0.18.2
gevent==21.12.0
google-ads==17.0.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-aiplatform==1.15.0
google-cloud-appengine-logging==1.1.2
google-cloud-audit-log==0.2.2
google-cloud-automl==2.7.3
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.6.2
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.2
google-cloud-build==3.8.3
google-cloud-container==2.10.8
google-cloud-core==2.3.1
google-cloud-datacatalog==3.8.1
google-cloud-dataplex==1.0.1
google-cloud-dataproc==4.0.3
google-cloud-dataproc-metastore==1.5.1
google-cloud-dlp==1.0.2
google-cloud-kms==2.11.2
google-cloud-language==1.3.2
google-cloud-logging==3.1.2
google-cloud-memcache==1.3.2
google-cloud-monitoring==2.9.2
google-cloud-orchestration-airflow==1.3.2
google-cloud-os-login==2.6.2
google-cloud-pubsub==2.13.0
google-cloud-redis==2.8.1
google-cloud-resource-manager==1.5.1
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==1.44.0
google-cloud-tasks==2.9.1
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.6.3
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
graphviz==0.20
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-gcp==0.2.2
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.6.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
looker-sdk==22.4.0
Mako==1.2.1
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.3
mdit-py-plugins==0.3.0
mdurl==0.1.1
netCDF4==1.6.0
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pandas-gbq==0.17.6
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.2.0
pluggy==1.0.0
prison==0.2.1
proto-plus==1.19.6
protobuf==3.20.0
psutil==5.9.1
psycopg2-binary==2.9.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
pyshp==2.3.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
setproctitle==1.2.3
Shapely==1.8.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlalchemy-bigquery==1.4.4
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
